Revert "XW-514 | rename CurrentUser to current-user for consistency"
This reverts commit 74b82380c99fb959f05739fdec9ab0acec834624.
