+++ front/scripts/main/CurrentUser.js
